[None][refactor] move Qwen-VL processor defaults fix to modeling_qwen2vl
Per review feedback, ``install_qwen_vl_processor_defaults_fix`` (and its
helpers ``_make_safe_get_num_multimodal_tokens``,
``_QWEN_VL_KWARGS_CLASS_BY_PROCESSOR``,
``_QWEN_VL_PROCESSOR_OUTPUT_KEYS``) is Qwen-VL-specific and does not
belong in the cross-model ``modeling_multimodal_utils`` surface. Move
the workaround into ``modeling_qwen2vl.py`` (next to its only
consumers) and re-import it from ``modeling_qwen3vl.py`` alongside the
existing ``Qwen2_5_VLVisionAttention`` import. Tests follow the new
import path. Behavior is unchanged.
