+++ tests/unittest/_torch/modeling/test_modeling_multimodal.py
@@ -17,9 +17,7 @@ from tensorrt_llm._torch.attention_backend.interface import AttentionRuntimeFeat
 from tensorrt_llm._torch.attention_backend.utils import get_attention_backend
 from tensorrt_llm._torch.metadata import KVCacheParams
 from tensorrt_llm._torch.model_config import ModelConfig
-from tensorrt_llm._torch.models.modeling_multimodal_utils import (
-    install_qwen_vl_processor_defaults_fix,
-)
+from tensorrt_llm._torch.models.modeling_qwen2vl import install_qwen_vl_processor_defaults_fix
 from tensorrt_llm._torch.pyexecutor.resource_manager import KVCacheManager
 from tensorrt_llm._utils import str_dtype_to_torch
 from tensorrt_llm.bindings.executor import KvCacheConfig
